+++ lib/cipi-api-sudoers.sh
@@ -75,6 +75,8 @@ www-data ALL=(root) NOPASSWD: /usr/local/bin/cipi app create *, \
                                /usr/local/bin/cipi service restart *, \
                                /usr/local/bin/cipi status, \
                                /usr/local/bin/cipi app webhook recreate *, \
+                               /usr/local/bin/cipi app fix-permissions, \
+                               /usr/local/bin/cipi app fix-permissions *, \
                                /usr/local/bin/cipi api ip-whitelist, \
                                /usr/local/bin/cipi api ip-whitelist *, \
                                /usr/local/bin/cipi-read-app-logs *, \